requirements.txt
PyMuPDF==1.26.3
pandas
python-multipart
Pillow
pytesseract
openai
fastapi
uvicorn
python-dotenv
requests
googlesearch-python 
requests 
trafilatura
openpyxl
groq
opencv-python